# 4-sqlite-performance: quick performance check