For me locally:

modernc.org/sqlite
```
writes: 9.681202343s
reads: 39.381931ms
```
mattn/go-sqlite3
```
writes: 6.531749771s
reads: 18.955969ms
```